Move definitions of ISubContainerResolution
